test(gyro_odometer): move the rewritten scenarios into the logic suite
The scratch file goes away and its contents arrive in the logic suite,
byte for byte. The only lines that are not part of the move are two
includes the moved text needs.
